+++ src/main/kotlin/com/github/victorrentea/livecoding/settings/AppSettingsComponent.kt
@@ -17,6 +17,8 @@ class AppSettingsComponent {
     private val staticImportsTextArea = JBTextArea(30, 10)
     private val showTestResultsSplashCheckbox = JBCheckBox()
     private val playTestResultsSoundCheckbox = JBCheckBox()
+    private val reportOpenFileToAddonCheckbox = JBCheckBox()
+    private val addonReportUrlField = JBTextField()
 
     var staticImports: List<String>
         get() = staticImportsTextArea.text.lines()
@@ -35,10 +37,23 @@ class AppSettingsComponent {
             playTestResultsSoundCheckbox.isSelected = newValue
         }
 
+    var reportOpenFileToAddon: Boolean
+        get() = reportOpenFileToAddonCheckbox.isSelected
+        set(newValue) {
+            reportOpenFileToAddonCheckbox.isSelected = newValue
+        }
+    var addonReportUrl: String
+        get() = addonReportUrlField.text
+        set(newValue) {
+            addonReportUrlField.text = newValue
+        }
+
     init {
         panel = FormBuilder.createFormBuilder()
             .addLabeledComponent(JBLabel("Show test results splash: "), showTestResultsSplashCheckbox)
             .addLabeledComponent(JBLabel("Play test results sounds: "), playTestResultsSoundCheckbox)
+            .addLabeledComponent(JBLabel("Report open file to add-on: "), reportOpenFileToAddonCheckbox)
+            .addLabeledComponent(JBLabel("Add-on report URL: "), addonReportUrlField)
             .addSeparator()
             .addLabeledComponentFillVertically("Methods or constants to auto-statically import:",
                 JBScrollPane(staticImportsTextArea))