+++ src/main/kotlin/com/github/victorrentea/livecoding/settings/SettingsConfigurable.kt
@@ -17,6 +17,8 @@ class SettingsConfigurable: Configurable {
         if (mySettingsComponent!!.staticImports != settings.staticImportsList) return true
         if (mySettingsComponent!!.showTestResultsSplash != settings.showTestResultsSplash) return true
         if (mySettingsComponent!!.playTestResultsSound != settings.playTestResultsSound) return true
+        if (mySettingsComponent!!.reportOpenFileToAddon != settings.reportOpenFileToAddon) return true
+        if (mySettingsComponent!!.addonReportUrl != settings.addonReportUrl) return true
         return false
     }
 
@@ -25,6 +27,8 @@ class SettingsConfigurable: Configurable {
         settings.staticImportsList = mySettingsComponent!!.staticImports
         settings.showTestResultsSplash = mySettingsComponent!!.showTestResultsSplash
         settings.playTestResultsSound = mySettingsComponent!!.playTestResultsSound
+        settings.reportOpenFileToAddon = mySettingsComponent!!.reportOpenFileToAddon
+        settings.addonReportUrl = mySettingsComponent!!.addonReportUrl
     }
 
     override fun getDisplayName() = "Live-Coding"
@@ -37,6 +41,8 @@ class SettingsConfigurable: Configurable {
         mySettingsComponent!!.staticImports = settings.staticImportsList
         mySettingsComponent!!.showTestResultsSplash = settings.showTestResultsSplash
         mySettingsComponent!!.playTestResultsSound = settings.playTestResultsSound
+        mySettingsComponent!!.reportOpenFileToAddon = settings.reportOpenFileToAddon
+        mySettingsComponent!!.addonReportUrl = settings.addonReportUrl
     }
     override fun getPreferredFocusedComponent(): JComponent {
         return mySettingsComponent!!.getPreferredFocusedComponent();